Reorganize codebase into routes/ and templates/ modules
- Split server.ts (1199 LOC) into src/routes/ with separate modules:
  - admin.ts - admin dashboard, events, settings, authentication
  - health.ts - health check endpoint
  - public.ts - home page and ticket reservation
  - setup.ts - initial system configuration
  - webhooks.ts - payment callbacks
  - middleware.ts - security headers, CORS, content-type validation
  - utils.ts - shared utilities (auth, cookies, responses)
  - types.ts - shared types

- Split html.ts (587 LOC) into src/templates/ with separate modules:
  - admin.ts - admin dashboard, event, settings pages
  - public.ts - home and ticket pages
  - payment.ts - payment success, cancel, error pages
  - setup.ts - initial setup pages
  - layout.ts - base layout and styles
  - fields.ts - form field definitions
  - csv.ts - CSV export generation

- Add path aliases #routes and #templates for clean imports
- Maintain backward compatibility through re-exports in server.ts and html.ts
- All 335 tests pass with 100% code coverage
