+++ package.json
@@ -9,6 +9,10 @@
     "#test-utils": "./src/test-utils/index.ts",
     "#test-utils/*": "./src/test-utils/*",
     "#lib/*": "./src/lib/*",
+    "#routes": "./src/routes/index.ts",
+    "#routes/*": "./src/routes/*",
+    "#templates": "./src/templates/index.ts",
+    "#templates/*": "./src/templates/*",
     "#src/*": "./src/*",
     "#test/*": "./test/*"
   },
