+++ src/routes/types.ts
@@ -1,0 +1,10 @@
+/**
+ * Shared types for route handlers
+ */
+
+/**
+ * Server context for accessing connection info
+ */
+export type ServerContext = {
+  requestIP?: (req: Request) => { address: string } | null;
+};
